source/aws-simulator/requirements.txt
CoAPthon==4.0.2
docopt==0.6.2
paho-mqtt==2.1.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.17.0
stomp.py==8.2.0
websocket-client==1.8.0
